Фикс отображения главной страницы в мобильной версии, добавлен шаблон мобильного контейнера главной страницы
