+++ src/components/Providers.tsx
@@ -27,7 +27,7 @@ export function Providers({ children }: { children: React.ReactNode }) {
         horizontal: "right",
       }}
     >
-      <div className="flex h-screen overflow-y-hidden bg-white dark:bg-gray-900">
+      <div className="flex flex-col md:flex-row h-screen overflow-y-hidden bg-white dark:bg-gray-900">
         {pathname !== "/login" && <AdaptiveNavbar />}
         <div className="flex-1 overflow-auto">{children}</div>
       </div>
