+++ src/components/ui/MainPageStatisticsMobileCard.tsx
@@ -1,0 +1,5 @@
+
+
+export const MainPageStatisticsMobileCard = ({ data }: any) => {
+    
+}
